[ros2] Extract Event to a new common ros package
