+++ ros2/src/common_interfaces/CMakeLists.txt
@@ -31,7 +31,6 @@ rosidl_generate_interfaces(${PROJECT_NAME}
   "srv/GetBool.srv"
   "srv/SetFloat32.srv"
   "srv/SetInt32.srv"
-
 )
 
 if(BUILD_TESTING)
