+++ ros2/src/detection/package.xml
@@ -17,6 +17,7 @@
   <build_depend>geometry_msgs</build_depend>
   <build_depend>tf2</build_depend>
   <build_depend>tf2_ros</build_depend>
+  <build_depend>common</build_depend>
   <build_depend>detection_interfaces</build_depend>
   <build_depend>camera_control</build_depend>
   <build_depend>cv_bridge</build_depend>
@@ -30,6 +31,7 @@
   <exec_depend>geometry_msgs</exec_depend>
   <exec_depend>tf2</exec_depend>
   <exec_depend>tf2_ros</exec_depend>
+  <exec_depend>common</exec_depend>
   <exec_depend>detection_interfaces</exec_depend>
   <exec_depend>camera_control</exec_depend>
   <exec_depend>cv_bridge</exec_depend>
