+++ ros2/src/runner_cutter_control/package.xml
@@ -12,6 +12,7 @@
   <build_depend>rclcpp</build_depend>
   <build_depend>std_msgs</build_depend>
   <build_depend>std_srvs</build_depend>
+  <build_depend>common</build_depend>
   <build_depend>runner_cutter_control_interfaces</build_depend>
   <build_depend>laser_control_interfaces</build_depend>
   <build_depend>camera_control_interfaces</build_depend>
@@ -21,6 +22,7 @@
   <exec_depend>rclcpp</exec_depend>
   <exec_depend>std_msgs</exec_depend>
   <exec_depend>std_srvs</exec_depend>
+  <exec_depend>common</exec_depend>
   <exec_depend>runner_cutter_control_interfaces</exec_depend>
   <exec_depend>laser_control_interfaces</exec_depend>
   <exec_depend>camera_control_interfaces</exec_depend>
