+++ ros2/src/runner_cutter_control/src/runner_cutter_control_node.cpp
@@ -1,11 +1,10 @@
 #include <fmt/core.h>
 
 #include <algorithm>
-#include <condition_variable>
-#include <mutex>
 
 #include "camera_control_interfaces/msg/device_state.hpp"
 #include "camera_control_interfaces/msg/state.hpp"
+#include "common/event.hpp"
 #include "detection_interfaces/msg/detection_result.hpp"
 #include "detection_interfaces/msg/detection_type.hpp"
 #include "laser_control_interfaces/msg/device_state.hpp"
@@ -34,42 +33,6 @@
 #include "std_srvs/srv/trigger.hpp"
 
 namespace {
-
-/**
- * Concurrency primitive that provides a shared flag that can be set and waited
- * on.
- */
-class Event {
- public:
-  Event() : flag_(false) {}
-
-  void set() {
-    std::lock_guard<std::mutex> lock(mtx_);
-    flag_ = true;
-    cv_.notify_all();
-  }
-
-  void clear() {
-    std::lock_guard<std::mutex> lock(mtx_);
-    flag_ = false;
-  }
-
-  void wait() {
-    std::unique_lock<std::mutex> lock(mtx_);
-    cv_.wait(lock, [this] { return flag_; });
-  }
-
-  bool wait_for(float timeoutSecs) {
-    std::unique_lock<std::mutex> lock(mtx_);
-    return cv_.wait_for(lock, std::chrono::duration<float>(timeoutSecs),
-                        [this] { return flag_; });
-  }
-
- private:
-  std::mutex mtx_;
-  std::condition_variable cv_;
-  bool flag_;
-};
 
 std::pair<int, int> millisecondsToRosTime(double milliseconds) {
   // ROS timestamps consist of two integers, one for seconds and one for
